Move lex::core module to separate file
